Sent vid/did -> arch mapping to a new file.
